+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+### Added
+- added symfony/event-dispatcher ^3.0.0 requirement
+- added optional feature to listen events from RequestSender
+
 ## [0.4.0] - 2017-08-09
 Fully refactored library, changed everything
 ### Added